+++ src/api/projects/releases.rs
@@ -10,6 +10,7 @@
 
 mod releases;
 
+pub use self::releases::ProjectReleaseOrderBy;
 pub use self::releases::ProjectReleases;
 pub use self::releases::ProjectReleasesBuilder;
 pub use self::releases::ProjectReleasesBuilderError;